check-validation-service/requirements.txt
strawberry-graphql[fastapi]
uvicorn